+++ src/boot/c/boot.c
@@ -6,543 +6,10 @@
 #include "boot.h"
 
 //
-//  Constructors
+//  Primitives
 //
 
-minim_object *make_char(int c) {
-    minim_char_object *o = GC_alloc(sizeof(minim_char_object));
-    o->type = MINIM_CHAR_TYPE;
-    o->value = c;
-    return ((minim_object *) o);
-}
-
-minim_object *make_symbol(const char *s) {
-    minim_symbol_object *o = GC_alloc(sizeof(minim_symbol_object));
-    int len = strlen(s);
-    
-    o->type = MINIM_SYMBOL_TYPE;
-    o->value = GC_alloc_atomic((len + 1) * sizeof(char));
-    strncpy(o->value, s, len + 1);
-    return ((minim_object *) o);
-}
-
-static void gc_port_dtor(void *ptr) {
-    minim_port_object *o = ((minim_port_object *) ptr);
-    if (minim_port_is_open(o))
-        fclose(minim_port(o));
-}
-
-minim_object *make_input_port(FILE *stream) {
-    minim_port_object *o = GC_alloc_opt(sizeof(minim_port_object), gc_port_dtor, NULL);
-    o->type = MINIM_PORT_TYPE;
-    o->flags = MINIM_PORT_READ_ONLY;
-    o->stream = stream;
-    return ((minim_object *) o);
-}
-
-minim_object *make_output_port(FILE *stream) {
-    minim_port_object *o = GC_alloc_opt(sizeof(minim_port_object), gc_port_dtor, NULL);
-    o->type = MINIM_PORT_TYPE;
-    o->flags = 0x0;
-    o->stream = stream;
-    return ((minim_object *) o);
-}
-
-minim_object *make_syntax(minim_object *e, minim_object *loc) {
-    minim_syntax_object *o = GC_alloc(sizeof(minim_syntax_object));
-    o->type = MINIM_SYNTAX_TYPE;
-    o->e = e;
-    o->loc = loc;
-    return ((minim_object *) o);
-}
-
-//
-//  System
-//
-
-void set_current_dir(const char *str) {
-    if (_set_current_dir(str) != 0) {
-        fprintf(stderr, "could not set current directory: %s\n", str);
-        exit(1);
-    }
-
-    current_directory(current_thread()) = make_string(str);
-}
-
-char *get_file_dir(const char *realpath) {
-    char *dirpath;
-    long i;
-
-    for (i = strlen(realpath) - 1; i >= 0 && realpath[i] != '/'; --i);
-    if (i < 0) {
-        fprintf(stderr, "could not resolve directory of path %s\n", realpath);
-        exit(1);
-    }
-
-    dirpath = GC_alloc_atomic((i + 2) * sizeof(char));
-    strncpy(dirpath, realpath, i + 1);
-    dirpath[i + 1] = '\0';
-    return dirpath;
-}
-
-minim_object *load_file(const char *fname) {
-    FILE *f;
-    minim_object *result, *expr;
-    char *old_cwd, *cwd;
-
-    f = fopen(fname, "r");
-    if (f == NULL) {
-        fprintf(stderr, "could not open file \"%s\"\n", fname);
-        exit(1);
-    }
-
-    old_cwd = _get_current_dir();
-    cwd = get_file_dir(_get_file_path(fname));
-    set_current_dir(cwd);
-
-    result = minim_void;
-    while ((expr = read_object(f)) != NULL)
-        result = eval_expr(expr, global_env(current_thread()));
-
-    set_current_dir(old_cwd);
-    fclose(f);
-    return result;
-}
-
-//
-//  Extra functions
-//
-
-// Recursively converts an object to syntax
-minim_object *to_syntax(minim_object *o) {
-    minim_object *it;
-
-    switch (o->type) {
-    case MINIM_SYNTAX_TYPE:
-        return o;
-    case MINIM_NULL_TYPE:
-    case MINIM_TRUE_TYPE:
-    case MINIM_FALSE_TYPE:
-    case MINIM_EOF_TYPE:
-    case MINIM_VOID_TYPE:
-    case MINIM_SYMBOL_TYPE:
-    case MINIM_FIXNUM_TYPE:
-    case MINIM_CHAR_TYPE:
-    case MINIM_STRING_TYPE:
-        return make_syntax(o, minim_false);
-    
-    case MINIM_PAIR_TYPE:
-        it = o;
-        do {
-            minim_car(it) = to_syntax(minim_car(it));
-            if (!minim_is_pair(minim_cdr(it))) {
-                if (!minim_is_null(minim_cdr(it)))
-                    minim_cdr(it) = to_syntax(minim_cdr(it));
-                return make_syntax(o, minim_false);
-            }
-            it = minim_cdr(it);
-        } while (1);
-    default:
-        fprintf(stderr, "datum->syntax: cannot convert to syntax");
-        exit(1);
-    }
-}
-
-minim_object *strip_syntax(minim_object *o) {
-    switch (o->type) {
-    case MINIM_SYNTAX_TYPE:
-        return strip_syntax(minim_syntax_e(o));
-    case MINIM_PAIR_TYPE:
-        return make_pair(strip_syntax(minim_car(o)), strip_syntax(minim_cdr(o)));
-    default:
-        return o;
-    }
-}
-
-//
-//  Primitive library
-//
-
-minim_object *is_symbol_proc(minim_object *args) {
-    // (-> any boolean)
-    return minim_is_symbol(minim_car(args)) ? minim_true : minim_false;
-}
-
-minim_object *is_char_proc(minim_object *args) {
-    // (-> any boolean)
-    return minim_is_char(minim_car(args)) ? minim_true : minim_false;
-}
-
-minim_object *is_input_port_proc(minim_object *args) {
-    // (-> any boolean)
-    return minim_is_input_port(minim_car(args)) ? minim_true : minim_false;
-}
-
-minim_object *is_output_port_proc(minim_object *args) {
-    // (-> any boolean)
-    return minim_is_output_port(minim_car(args)) ? minim_true : minim_false;
-}
-
-minim_object *is_syntax_proc(minim_object *args) {
-    // (-> any boolean)
-    return minim_is_syntax(minim_car(args)) ? minim_true : minim_false;
-}
-
-minim_object *char_to_integer_proc(minim_object *args) {
-    // (-> char integer)
-    minim_object *o = minim_car(args);
-    if (!minim_is_char(o))
-        bad_type_exn("char->integer", "char?", o);
-    return make_fixnum(minim_char(o));
-}
-
-minim_object *integer_to_char_proc(minim_object *args) {
-    // (-> integer char)
-    minim_object *o = minim_car(args);
-    if (!minim_is_fixnum(o))
-        bad_type_exn("integer->char", "integer?", o);
-    return make_char(minim_fixnum(minim_car(args)));
-}
-
-minim_object *syntax_e_proc(minim_object *args) {
-    // (-> syntax any)
-    if (!minim_is_syntax(minim_car(args)))
-        bad_type_exn("syntax-e", "syntax?", minim_car(args));
-    return minim_syntax_e(minim_car(args));
-}
-
-minim_object *syntax_loc_proc(minim_object *args) {
-    // (-> syntax any)
-    if (!minim_is_syntax(minim_car(args)))
-        bad_type_exn("syntax-loc", "syntax?", minim_car(args));
-    return minim_syntax_loc(minim_car(args));
-}
-
-minim_object *to_syntax_proc(minim_object *args) {
-    // (-> any syntax)
-    return to_syntax(minim_car(args));
-}
-
-minim_object *current_input_port_proc(minim_object *args) {
-    // (-> input-port)
-    return input_port(current_thread());
-}
-
-minim_object *current_output_port_proc(minim_object *args) {
-    // (-> output-port)
-    return output_port(current_thread());
-}
-
-minim_object *open_input_port_proc(minim_object *args) {
-    // (-> string input-port)
-    minim_object *port;
-    FILE *stream;
-
-    if (!minim_is_string(minim_car(args)))
-        bad_type_exn("open-input-port", "string?", minim_car(args));
-
-    stream = fopen(minim_string(minim_car(args)), "r");
-    if (stream == NULL) {
-        fprintf(stderr, "could not open file \"%s\"\n", minim_string(minim_car(args)));
-        exit(1);
-    }
-
-    port = make_input_port(stream);
-    minim_port_set(port, MINIM_PORT_OPEN);
-    return port;
-}
-
-minim_object *open_output_port_proc(minim_object *args) {
-    // (-> string output-port)
-    minim_object *port;
-    FILE *stream;
-
-    if (!minim_is_string(minim_car(args)))
-        bad_type_exn("open-output-port", "string?", minim_car(args));
-    
-    stream = fopen(minim_string(minim_car(args)), "w");
-    if (stream == NULL) {
-        fprintf(stderr, "could not open file \"%s\"\n", minim_string(minim_car(args)));
-        exit(1);
-    }
-
-    port = make_output_port(stream);
-    minim_port_set(port, MINIM_PORT_OPEN);
-    return port;
-}
-
-minim_object *close_input_port_proc(minim_object *args) {
-    // (-> input-port)
-    if (!minim_is_input_port(minim_car(args)))
-        bad_type_exn("close-input-port", "input-port?", minim_car(args));
-
-    fclose(minim_port(minim_car(args)));
-    minim_port_unset(minim_car(args), MINIM_PORT_OPEN);
-    return minim_void;
-}
-
-minim_object *close_output_port_proc(minim_object *args) {
-    // (-> output-port)
-    if (!minim_is_output_port(minim_car(args)))
-        bad_type_exn("close-output-port", "output-port?", minim_car(args));
-
-    fclose(minim_port(minim_car(args)));
-    minim_port_unset(minim_car(args), MINIM_PORT_OPEN);
-    return minim_void;
-}
-
-minim_object *read_proc(minim_object *args) {
-    // (-> any)
-    // (-> input-port any)
-    minim_object *in_p, *o;
-
-    if (minim_is_null(args)) {
-        in_p = input_port(current_thread());
-    } else {
-        in_p = minim_car(args);
-        if (!minim_is_input_port(in_p))
-            bad_type_exn("read", "input-port?", in_p);
-    }
-
-    o = read_object(minim_port(in_p));
-    return (o == NULL) ? minim_eof : o;
-}
-
-minim_object *read_char_proc(minim_object *args) {
-    // (-> char)
-    // (-> input-port char)
-    minim_object *in_p;
-    int ch;
-    
-    if (minim_is_null(args)) {
-        in_p = input_port(current_thread());
-    } else {
-        in_p = minim_car(args);
-        if (!minim_is_input_port(in_p))
-            bad_type_exn("read-char", "input-port?", in_p);
-    }
-
-    ch = getc(minim_port(in_p));
-    return (ch == EOF) ? minim_eof : make_char(ch);
-}
-
-minim_object *peek_char_proc(minim_object *args) {
-    // (-> char)
-    // (-> input-port char)
-    minim_object *in_p;
-    int ch;
-    
-    if (minim_is_null(args)) {
-        in_p = input_port(current_thread());
-    } else {
-        in_p = minim_car(args);
-        if (!minim_is_input_port(in_p))
-            bad_type_exn("peek-char", "input-port?", in_p);
-    }
-
-    ch = getc(minim_port(in_p));
-    ungetc(ch, minim_port(in_p));
-    return (ch == EOF) ? minim_eof : make_char(ch);
-}
-
-minim_object *char_is_ready_proc(minim_object *args) {
-    // (-> boolean)
-    // (-> input-port boolean)
-    minim_object *in_p;
-    int ch;
-
-    if (minim_is_null(args)) {
-        in_p = input_port(current_thread());
-    } else {
-        in_p = minim_car(args);
-        if (!minim_is_input_port(in_p))
-            bad_type_exn("peek-char", "input-port?", in_p);
-    }
-
-    ch = getc(minim_port(in_p));
-    ungetc(ch, minim_port(in_p));
-    return (ch == EOF) ? minim_false : minim_true;
-}
-
-minim_object *display_proc(minim_object *args) {
-    // (-> any void)
-    // (-> any output-port void)
-    minim_object *out_p, *o;
-
-    o = minim_car(args);
-    if (minim_is_null(minim_cdr(args))) {
-        out_p = output_port(current_thread());
-    } else {
-        out_p = minim_cadr(args);
-        if (!minim_is_output_port(out_p))
-            bad_type_exn("display", "output-port?", out_p);
-    }
-
-    write_object2(minim_port(out_p), o, 0, 1);
-    return minim_void;
-}
-
-minim_object *write_proc(minim_object *args) {
-    // (-> any void)
-    // (-> any output-port void)
-    minim_object *out_p, *o;
-
-    o = minim_car(args);
-    if (minim_is_null(minim_cdr(args))) {
-        out_p = output_port(current_thread());
-    } else {
-        out_p = minim_cadr(args);
-        if (!minim_is_output_port(out_p))
-            bad_type_exn("display", "output-port?", out_p);
-    }
-
-    write_object(minim_port(out_p), o);
-    return minim_void;
-}
-
-minim_object *write_char_proc(minim_object *args) {
-    // (-> char void)
-    // (-> char output-port void)
-    minim_object *out_p, *ch;
-
-    ch = minim_car(args);
-    if (!minim_is_char(ch))
-        bad_type_exn("write-char", "char?", ch);
-
-    if (minim_is_null(minim_cdr(args))) {
-        out_p = output_port(current_thread());
-    } else {
-        out_p = minim_cadr(args);
-        if (!minim_is_output_port(out_p))
-            bad_type_exn("display", "output-port?", out_p);
-    }
-
-    putc(minim_char(ch), minim_port(out_p));
-    return minim_void;
-}
-
-minim_object *newline_proc(minim_object *args) {
-    // (-> void)
-    // (-> output-port void)
-    minim_object *out_p;
-
-    if (minim_is_null(args)) {
-        out_p = output_port(current_thread());
-    } else {
-        out_p = minim_car(args);
-        if (!minim_is_output_port(out_p))
-            bad_type_exn("display", "output-port?", out_p);
-    }
-
-    putc('\n', minim_port(out_p));
-    return minim_void;
-}
-
-minim_object *load_proc(minim_object *args) {
-    // (-> string any)
-    if (!minim_is_string(minim_car(args)))
-        bad_type_exn("load", "string?", minim_car(args));
-    return load_file(minim_string(minim_car(args)));
-}
-
-minim_object *error_proc(minim_object *args) {
-    // (-> (or #f string symbol)
-    //     string?
-    //     any ...
-    //     any)
-    minim_object *who, *message, *reasons;
-
-    who = minim_car(args);
-    if (!minim_is_false(who) && !minim_is_symbol(who) && !minim_is_string(who))
-        bad_type_exn("error", "(or #f string? symbol?)", who);
-
-    message = minim_cadr(args);
-    if (!minim_is_string(message))
-        bad_type_exn("error", "string?", message);
-    
-    if (minim_is_false(who))
-        printf("error");
-    else
-        write_object2(stderr, who, 1, 1);
-
-    fprintf(stderr, ": ");
-    write_object2(stderr, message, 1, 1);
-    fprintf(stderr, "\n");
-
-    reasons = minim_cddr(args);
-    while (!minim_is_null(reasons)) {
-        fprintf(stderr, " ");
-        write_object2(stderr, minim_car(reasons), 1, 1);
-        fprintf(stderr, "\n");
-        reasons = minim_cdr(reasons);
-    }
-
-    exit(1);
-}
-
-minim_object *exit_proc(minim_object *args) {
-    // (-> any)
-    exit(0);
-}
-
-minim_object *syntax_error_proc(minim_object *args) {
-    // (-> symbol string any)
-    // (-> symbol string syntax any)
-    // (-> symbol string syntax syntaxs any)
-    minim_object *what, *why, *where, *sub;
-
-    what = minim_car(args);
-    why = minim_cadr(args);
-
-    if (!minim_is_symbol(what))
-        bad_type_exn("syntax-error", "symbol?", what);
-    if (!minim_is_string(why))
-        bad_type_exn("syntax-error", "string?", why);
-
-    fprintf(stderr, "%s: %s\n", minim_symbol(what), minim_string(why));
-    if (!minim_is_null(minim_cddr(args))) {
-        where = minim_car(minim_cddr(args));
-        if (!minim_is_syntax(where))
-            bad_type_exn("syntax-error", "syntax?", where);
-
-        if (!minim_is_null(minim_cdr(minim_cddr(args)))) {
-            sub = minim_cadr(minim_cddr(args));
-            if (!minim_is_syntax(sub))
-                bad_type_exn("syntax-error", "syntax?", sub);
-
-            fputs("  at: ", stderr);
-            write_object2(stderr, sub, 1, 1);
-            fprintf(stderr, "\n");
-        }
-
-        fputs("  in: ", stderr);
-        write_object2(stderr, where, 1, 1);
-        fprintf(stderr, "\n");
-    }
-
-    exit(1);
-}
-
-minim_object *current_directory_proc(minim_object *args) {
-    // (-> string)
-    // (-> string void)
-    minim_object *path;
-
-    if (minim_is_null(args)) {
-        // getter
-        return current_directory(current_thread());
-    } else {
-        // setter
-        path = minim_car(args);
-        if (!minim_is_string(path))
-            bad_type_exn("current-directory", "string?", path);
-        
-        set_current_dir(minim_string(path));
-        return minim_void;
-    }
-}
-
-minim_object *boot_expander_proc(minim_object *args) {
+static minim_object *boot_expander_proc(minim_object *args) {
     // (-> boolean)
     // (-> boolean void)
     minim_object *val;
